server/requirements.txt
backcall==0.2.0
blis==0.7.4
catalogue==2.0.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
contextvars==2.4
cryptography==3.4.6
cycler==0.10.0
cymem==2.0.5
Cython==0.29.22
dataclasses==0.8
decorator==4.4.2
dnspython==2.1.0
email-validator==1.1.2
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.0.0/en_core_web_lg-3.0.0-py3-none-any.whl
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0-py3-none-any.whl
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.10
Flask-WTF==0.14.3
gensim==3.8.3
idna==2.10
immutables==0.15
importlib-metadata==3.7.0
ipython==7.16.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
kiwisolver==1.3.1
llvmlite==0.34.0
MarkupSafe==1.1.1
matplotlib==3.3.4
murmurhash==1.0.5
nltk==3.5
numba==0.51.2
numpy==1.19.5
openTSNE==0.5.1
packaging==20.9
pandas==1.1.5
parso==0.8.1
pathy==0.4.0
pdfminer.six==20201018
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
preshed==3.0.5
prompt-toolkit==3.0.16
ptyprocess==0.7.0
pycparser==2.20
pydantic==1.7.3
Pygments==2.8.0
pymongo==3.11.3
pynndescent==0.5.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
regex==2020.11.13
requests==2.25.1
sacremoses==0.0.43
scikit-learn==0.24.1
scipy==1.5.4
seaborn==0.11.1
sentence-transformers==0.4.1.2
sentencepiece==0.1.95
six==1.15.0
smart-open==3.0.0
sortedcontainers==2.3.0
spacy==3.0.3
spacy-legacy==3.0.1
srsly==2.4.0
thinc==8.0.1
threadpoolctl==2.1.0
tokenizers==0.10.1
torch==1.7.1
tqdm==4.57.0
traitlets==4.3.3
transformers==4.3.2
typer==0.3.2
typing-extensions==3.7.4.3
umap-learn==0.5.1
urllib3==1.26.3
wasabi==0.8.2
wcwidth==0.2.5
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.4.0
